--- src/lib/components/portal/index.ts
@@ -1,1 +1,0 @@
-export {default as Root} from './root.svelte';
