+++ src/lib/components/range-date-picker/index.ts
@@ -1,1 +1,2 @@
+export {default as Portal} from '../portal.svelte';
 export {default as Root} from './root.svelte';
