+++ src/lib/index.ts
@@ -1,6 +1,5 @@
 export * as Calendar from './components/calendar/index.js';
 export * as DatePicker from './components/date-picker/index.js';
-export * as Portal from './components/portal/root.svelte';
 export * as RangeCalendar from './components/range-calendar/index.js';
 export * as RangeDatePicker from './components/range-date-picker/index.js';
